requirements.txt
pandas
lark-parser
numpy